BF-Systems: New base classes for state based systems #549
