+++ src/mlpro/rl/models.py
@@ -7,6 +7,7 @@
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2021-10-26  1.0.0     DA       Splitted modules for sar-, env-, agent- and training classes.
 ## --                                Further change logs are documented there.
+## -- 2022-11-29  1.1.0     DA       Refactoring
 ## -------------------------------------------------------------------------------------------------
 
 """
@@ -14,7 +15,7 @@ This module provides model classes for reinforcement learning tasks. See sub-mdu
 information.
 """
 
-from mlpro.rl.models_sar import *
 from mlpro.rl.models_env import *
+from mlpro.rl.models_env_ada import *
 from mlpro.rl.models_agents import *
 from mlpro.rl.models_train import *
